project: rename mac to macx
The platform name "macx" is only for Mac OS X, however, "mac" is for both OSX and iOS.
And the settings in "mac" will overwrite the settings in "ios"
After corrected this platform setting, I'm sure that "ios" use the right Info.plist,
so we don't need that "NOTE" description in README.

   $ plutil -convert xml1 Debug-iphoneos/Terrarium.app/Info.plist
   $ grep -A1 CFBundleInfoDictionaryVersion Debug-iphoneos/Terrarium.app/Info.plist
	<key>CFBundleInfoDictionaryVersion</key>
	<string>6.0</string>

But this issue only happens when submitting app to App Store, so I didn't test it.
If I'm wrong, please reverst this commit.

Signed-off-by: AceLan Kao <acelan@acelan.idv.tw>